feat: added .env and seperated firebase creds
