+++ .gitignore
@@ -29,3 +29,4 @@ ui.debug.log
 /emulator-data
 
 /.vscode
+/.env
